Rename instruments -> wrappers
